Added a models.py file
It's based on the database diagram I put in the discord server, however I made a few changes as I went realizing better ways to do stuff. For example the TaskSchedule table is new since django doesn't like having lists as a datatype for a table entry... Breaking it out into its own table with a 1-1 relation is better practice anyways.